+++ src/migrations/m180826_000000_add_comment_date_column.php
@@ -11,7 +11,9 @@ class m180826_000000_add_comment_date_column extends Migration
     public function safeUp()
     {   
         // First create the new column
-        $this->addColumn('{{%comments_comments}}', 'commentDate', $this->dateTime()->after('userAgent')->notNull());
+        if (!$this->db->columnExists('{{%comments_comments}}', 'commentDate')) {
+            $this->addColumn('{{%comments_comments}}', 'commentDate', $this->dateTime()->after('userAgent')->notNull());
+        }
 
         // For existing records, backfill the new column with the existing values from dateCreated
         $this->update('{{%comments_comments}}', ['commentDate' => new Expression('dateCreated')]);