+++ src/records/Comment.php
@@ -3,6 +3,7 @@ namespace verbb\comments\records;
 
 use craft\db\ActiveRecord;
 use craft\records\Element;
+use craft\records\Site;
 use craft\records\User;
 
 use yii\db\ActiveQueryInterface;
@@ -25,6 +26,11 @@ class Comment extends ActiveRecord
     public function getOwner(): ActiveQueryInterface
     {
         return $this->hasOne(Element::class, ['id' => 'ownerId']);
+    }
+
+    public function getOwnerSite(): ActiveQueryInterface
+    {
+        return $this->hasOne(Site::class, ['id' => 'ownerSiteId']);
     }
 
     public function getUser(): ActiveQueryInterface