+++ src/services/CommentsService.php
@@ -43,10 +43,14 @@ class CommentsService extends Component
         $templatePath = $this->_getTemplatePath();
         $view->setTemplatesPath($templatePath);
 
-        $query = $this->fetch($criteria);
+        $query = Comment::find();
         $query->ownerId($elementId);
         $query->level('1');
         $query->orderBy('commentDate desc');
+
+        if ($criteria) {
+            Craft::configure($query, $criteria);
+        }
 
         $element = Craft::$app->getElements()->getElementById($elementId);
         $id = 'cc-w-' . $elementId;
@@ -55,6 +59,8 @@ class CommentsService extends Component
             'baseUrl' => UrlHelper::actionUrl(),
             'csrfTokenName' => Craft::$app->getConfig()->getGeneral()->csrfTokenName,
             'csrfToken' => Craft::$app->getRequest()->getCsrfToken(),
+            'recaptchaEnabled' => (bool)$settings->recaptchaEnabled,
+            'recaptchaKey' => $settings->recaptchaKey,
             'translations' => [
                 'reply' => Craft::t('comments', 'Reply'),
                 'close' => Craft::t('comments', 'Close'),